requirements.txt
groq==0.13.1
pydantic==2.10.4
python-dotenv==1.0.1
langchain~=0.3.13
langchain-core~=0.3.28
langchain-community~=0.3.13
yt-dlp~=2024.12.13